Nightly update

+++ Source/Kernel/MemoryManager/PageDirectory.h
@@ -108,7 +108,7 @@ public:
     
     //void faultHandler();
     inline static ulong getPhysAddress(ulong address) { return address & ~0xFFFFFFFFC0000000; }
-    inline ulong getPhysicalAddr() { return getPhysAddress(pages); }
+    inline ulong getPhysicalAddr() { return getPhysAddress((ulong)pages); }
 };
 
 #endif
